% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_function.R
\name{pubmed_mesh_online}
\alias{pubmed_mesh_online}
\title{获取 PubMed MeSH 主题词和条目词}
\usage{
pubmed_mesh_online(csv_file)
}
\arguments{
\item{csv_file}{字符串，包含输入数据的 CSV 文件路径，CSV 文件应该包含关键词列。}
}
\value{
返回一个列表，包含 MeSH 主题词和条目词的向量。
}
\description{
这个函数从 PubMed 获取指定关键词的 MeSH 主题词和条目词，并返回它们。
}
\examples{
# 假设有一个名为 "keywords.csv" 的文件，其中包含关键词
pubmed_mesh_online("path_to_your_csv_file.csv")

}
